SWORD25: Move all lua serialization helper functions to their own file
